Move HorizontalTextsWithMaximumSpaceTaken to core-ui
